Renamed Juniper Admin into Juniper client and abstracted it out into its own class
